requirements.txt
langchain==0.3.13
langchain-core==0.3.28
langchain-groq==0.2.1
langchain-community==0.3.13
chromadb==0.4.22
pandas==2.3.3
numpy==1.26.4
scikit-learn==1.4.0
scipy==1.16.3
matplotlib==3.8.2
seaborn==0.13.1
streamlit==1.52.2
python-dotenv==1.0.0
sentence-transformers==2.2.2